+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.25.0] - 2021-07-28
+## [3.26.0] - 2021-07-29
+
+### Added
+
+- Retry Imaging Set Errors methods for IImagingJobService. - [RTF-1368](https://jira.kcura.com/browse/RTF-1368)
+- Update Imaging Job Priority methods for IImagingJobService. - [RTF-1369](https://jira.kcura.com/browse/RTF-1369)
+
+## [3.25.0] - 2021-07-29
 
 ### Added